% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_price_Zoopla.R
\name{get_price_Zoopla}
\alias{get_price_Zoopla}
\title{Get the price estimate of a UK house from Zoopla}
\usage{
get_price_Zoopla(link_house)
}
\arguments{
\item{link_house}{The link to the relevant page for the property (rather than street) on Zoopla.}
}
\description{
This function scrapes a Zoopla page and returns the associated estimated price of the property.
}
\examples{
\dontrun{
get_price_Zoopla(link_house = 'https://www.zoopla.co.uk/property/1-oak-road/reigate/rh2-0bp/19534314')
get_price_Zoopla(link_house = 'https://www.zoopla.co.uk/property/10-waverleigh-road/cranleigh/gu6-8bz/9852441')
}
}
\keyword{Zoopla}
