% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_non.R
\name{stop_not_positive}
\alias{stop_not_positive}
\title{Stop the function the value entered is (or is below) zero}
\usage{
stop_not_positive(value_entered)
}
\arguments{
\item{value_entered}{The value to be error-checked by the function}
}
\description{
Error function to stop the function if the value entered is (or is below) zero
}
\examples{
\dontrun{
stop_not_positive(value_entered = age)
}

}
\keyword{errors}
