% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_non.R
\name{stop_not_scalar_double}
\alias{stop_not_scalar_double}
\title{Stop the function if zero or more than one value is entered}
\usage{
stop_not_scalar_double(value_entered)
}
\arguments{
\item{value_entered}{The value to be error-checked by the function}
}
\description{
Error function to stop the function if zero or more than one value is entered
}
\examples{
\dontrun{
stop_not_scalar(value_entered = age)
}

}
\keyword{errors}
