% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tax_parameters}
\alias{tax_parameters}
\title{Key income tax parameters for the UK.}
\format{
A data frame with 3 rows (one for each tax year) and 14 variables:
\describe{
  \item{year_tax_end}{the calendar year in which the end of the tax year occurs}
  \item{level_ni_lower}{the lower level breakpoint for National Insurance}
  \item{level_ni_upper}{the upper level breakpoint for National Insurance}
  \item{level_allowance_lower}{the lower level breakpoint for personal allowance}
  \item{level_allowance_upper}{the upper level breakpoint for personal allowance}
  \item{level_tax_higher}{the income breakpoint for higher rate income tax}
  \item{level_tax_upper}{the income breakpoint for upper rate income tax}
  \item{rate_ni_lower}{the lower level tax rate for National Insurance}
  \item{rate_ni_upper}{the upper level tax rate for National Insurance}
  \item{rate_allowance_drop}{the rate at which personal allowance drops after `level_allowance_upper`}
  \item{rate_tax_basic}{the basic rate of income tax}
  \item{rate_tax_higher}{the higher rate of income tax}
  \item{rate_tax_upper}{the upper rate of income tax}
  \item{rate_tax_sales}{the rate of sales tax, also called VAT}
}
}
\source{
\url{https://www.crunch.co.uk/knowledge-tax/tax-rates-thresholds-and-allowances-for-current-tax-year/}
}
\usage{
tax_parameters
}
\description{
A dataset containing the main tax parameters used for calculating net income for the UK.
}
\keyword{datasets}
