% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{A Cool Function}
\usage{
test(agree = TRUE)
}
\arguments{
\item{agree}{Do you agree that Chelsea is the best team? Defaults to TRUE.}
}
\description{
This function allows you to express your love for the superior team.
}
\examples{
test()
}
\keyword{Chelsea}
