% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p0bservations-package.R
\docType{package}
\name{p0bservations-package}
\alias{p0bservations}
\alias{p0bservations-package}
\title{p0bservations: Assorted functions and observations by p0bs}
\description{
Provides assorted functions by p0bs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/p0bs/p0bservations}
}

}
\author{
\strong{Maintainer}: Robin Penfold \email{robinpenfold@icloud.com}

}
\keyword{internal}
