% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_ruin_rate.R
\name{probability_ruin_rate}
\alias{probability_ruin_rate}
\title{Calculate the spend rate for a given probability of retirement ruin}
\usage{
probability_ruin_rate(
  return_expected,
  return_sd,
  life_remaining_expected,
  value_probability_ruin
)
}
\arguments{
\item{return_expected}{The expected real return of the entire pension portfolio}

\item{return_sd}{The projected standard deviation of the returns of the entire pension portfolio}

\item{life_remaining_expected}{The median projected remaining lifespan of the individual in question}

\item{value_probability_ruin}{The desired probability of retirement ruin (which is used to solve for the corresponding spend rate)}
}
\description{
This function uses the Milevsky-Robinson to analyse the spend rate for a given probability of retirement ruin, by parsimoniously meshing investment risk and return, mortality estimates and spending rates without resorting to opaque Monte Carlo simulations. For further details, see: Milevsky, M. and C. Robinson; "A Sustainable Spending Rate without Simulation"; Financial Analysts Journal, Volume 61, Number 6. (2005). Please note that these are approximations, so do not rely on them for financial returns or planning.
}
\examples{
probability_ruin_rate(
  return_expected = 0.07, 
  return_sd = 0.2, 
  life_remaining_expected = 28.1, 
  value_probability_ruin = 0.1
  )

}
\keyword{Retirement}
